core: move tee_ta_manager.c out of arch/arm
Moves most parts of tee_ta_manager.c into core/kernel, keeping only
tee_ta_verify_param() in the original file as it uses architecture
specific checks.

Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU)
