+++ core/kernel/sub.mk
@@ -1,5 +1,6 @@
 srcs-y += assert.c
 srcs-y += tee_dispatch.c
+srcs-y += tee_ta_manager.c
 srcs-y += tee_misc.c
 srcs-y += panic.c
 srcs-y += handle.c
